fiks 500-feil i forhåndsvisning ved åpning av panel
