+++ package.json
@@ -5,7 +5,7 @@
     "engines": {
         "node": "24.x"
     },
-    "packageManager": "pnpm@11.1.2+sha512.415a1cc25974731e75455c1468371be74c5aa5fb7621b50d4056d222451609f11412f23fd602e6169f1e060466641f798597e1be961a10688836a67b16569499",
+    "packageManager": "pnpm@11.5.1+sha512.93f7b57422ea7068257235b4c16eb60762eb68e1dc23723199cc739043ea9be2c4143274a399d8c6defa2b1176226d9ca1c4b63482d6200c1a8fbaa78c1d1485",
     "scripts": {
         "build": "pnpm run build:server && pnpm run build:nextjs",
         "build:nextjs": "pnpm --filter ./packages/nextjs run build",
