+++ packages/nextjs/src/components/_editor-only/redaktorvarsler/harRedaktorFeil.tsx
@@ -1,6 +1,5 @@
 import { ContentProps } from 'types/content-props/_content-common';
 import { KortUrlVarsel } from './varsler/kort-url/KortUrlVarsel';
-import { DuplikateIder } from './varsler/duplikate-ider/DuplikateIder';
 import { SkjemanummerVarsel } from './varsler/skjemanummer/SkjemanummerVarsel';
 import { KontaktinformasjonVarsel } from './varsler/kontaktinformasjon/KontaktinformasjonVarsel';
 import { FormatertInnholdUtenforInnholdsseksjon } from './varsler/formatert-innhold-utenfor-innholdsseksjon/FormatertInnholdUtenforInnholdsseksjon';
@@ -13,13 +12,12 @@ export const harRedaktorfeil = (content: ContentProps): boolean => {
     if (!isEditorView) {
         return false;
     }
-    return (
-        KortUrlVarsel({ content }) !== null ||
-        DuplikateIder({}) !== null ||
-        SkjemanummerVarsel({ content }) !== null ||
-        KontaktinformasjonVarsel({ content }) !== null ||
-        FormatertInnholdUtenforInnholdsseksjon({ content }) !== null ||
-        FragmentUtenforInnholdsseksjon({ content }) !== null ||
-        HtmlAreaInnholderDiv({ content }) !== null
-    );
+    return [
+        KortUrlVarsel({ content }),
+        SkjemanummerVarsel({ content }),
+        KontaktinformasjonVarsel({ content }),
+        FormatertInnholdUtenforInnholdsseksjon({ content }),
+        FragmentUtenforInnholdsseksjon({ content }),
+        HtmlAreaInnholderDiv({ content }),
+    ].some((result) => result !== null);
 };
